+++ project/stampee/src/main/java/session/MemberSession.java
@@ -1,25 +1,26 @@
-package domain;
+package session;
 
 import static exception.ErrorMessage.*;
 
 import dto.response.LoggedMemberDto;
 
-public class LoginSession {
-	private static LoginSession instance;
+public class MemberSession {
+	private static MemberSession instance;
 	private final LoggedMemberDto loggedMemberDto;
 
-	public LoginSession(LoggedMemberDto loggedMemberDto) {
+	public MemberSession(LoggedMemberDto loggedMemberDto) {
 		this.loggedMemberDto = loggedMemberDto;
 	}
-	//로그인 시 새로운 세션을 생성
-	public static synchronized LoginSession getInstance(LoggedMemberDto loggedMemberDto){
+
+	public static synchronized MemberSession getInstance(LoggedMemberDto loggedMemberDto){
 		if(instance == null){
-			instance = new LoginSession(loggedMemberDto);
+			instance = new MemberSession(loggedMemberDto);
+			return instance;
 		}
 		return instance;
 	}
-	//세션이 있는 경우 반환, 없는 경우 예외 발생
-	public static  synchronized LoginSession getInstance(){
+
+	public static  synchronized MemberSession getInstance(){
 		if(instance == null){
 			throw new IllegalArgumentException(USER_NOT_LOGGED_IN.getErrorMessage());
 		}
